"Restrukturert pakker, slettet unødig DTO
A comprehensive refactor is performed to make package structure more logical. Files related to the database are now in a 'db' package, and request objects are in a 'request' package, aligning with conventional package layout. The 'DokumentEventDto' was identified as obsolete, thus deleted. Access modifiers of some classes and interfaces have also been adjusted to maintain the appropriate level of encapsulation."